Changes as per feedback
We do not want a full screen website. The maximum width should be
1000px and it should be fix size since it has not been made responsive.
 - There is confusion between this and last point. I think 1288px width
is looking better as 1000px will have side padding.
    And with 1000px we will have to scale the whole website again.
Please confirm if I need to fix the size to 1288px .

Images have not been optimised. Images are very heavy in size.
 - Will optimize the images for fast rendering.

Carousel text should be part of image.
- Will update it.

Sub menu (notice board, register, contact) too much padding on top.
- In mock up there was more padding on top, so I kept more padding
intentionally. Will fix it.

Committee Members page is empty.
- As discussed with you this data might not be static, so I left it
blank intentionally.

News Events text background not properly visible.
- Will decrease the transparency for darker background.

Font is different. That is ok. But font sizes are not as per
illustration. Use sans-serif font since that is standardly available
and sizes can be set on the basis of that.
- As discussed the fonts which are used in the mock up are
HelveticaNeueLT-Roman and HelveticaNeueLT-Thin which are paid and so
not easliy available.
  Please confirm if we are good with sans-serif font.

When i resize to 1288px and set font as sans-serif the page looks
proper. but there is a gap on the right side since page content area is
smaller than the top green header.
- Will fix the right gap issue. And as mentioned in point 1 please
confirm if I should fix the width to 1288px.